requirements.txt
PyPDF2
openai
python-dotenv
flask
flask-cors
